refactor: функция генерации фильтров перенесена в модель - исправлены наименования view
