+++ src/filter.js
@@ -1,4 +1,4 @@
-import { FilterType } from './const';
+import { FilterType } from './const.js';
 import dayjs from 'dayjs';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
 import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
